(frontend)cleaning code and smile fix
fix: vendor delete option to choose type product

+++ app/Http/Controllers/Backend/ChildCategoryController.php
@@ -6,6 +6,8 @@ use App\DataTables\ChildCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\ChildCategory;
+use App\Models\HomePageSettings;
+use App\Models\Product;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
@@ -106,6 +108,19 @@ class ChildCategoryController extends Controller
     public function destroy(string $id)
     {
         $childCategory = childCategory::findOrFail($id);
+        if(Product::where('child_category_id', $childCategory->id)->count() > 0){
+            return response(['status' => 'error', 'message' =>'Kategoria posiada jest połączony z produktem']);
+        }
+        $homeSettings = HomePageSettings::all();
+
+        foreach($homeSettings as $item){
+            $array = json_decode($item->value, true);
+            $collection = collect($array);
+            if($collection->contains('child_category', $childCategory->id)){
+                return response(['status' => 'error', 'message' => 'Kategoria posiada jest połączony z produktem']);
+            }
+        }
+       
         $childCategory->delete();
 
         return response(['status' => 'success', 'message' =>'Usunięto pod pod kategorie']);
